Rename refactored "jupyterlab-go-to-definition" to "code-jumpers"
