+++ packages/code-jumpers/README.md
@@ -1,0 +1,3 @@
+# Code Jumpers
+
+Implementation underlying the jump to definition functionality in [JupyterLab-LSP](https://github.com/krassowski/jupyterlab-lsp).
